# MR: dev → master

## 1. Migration: Maven → Gradle
- Replaced all `pom.xml` with a multi-module Gradle build (`settings.gradle.kts`, `build.gradle.kts`, `buildSrc/java-conventions`, `gradle/libs.versions.toml` version catalog + Gradle wrapper).
- Added a Java CI workflow (`.github/workflows`) and removed all Maven references.
- Updated `AGENTS.md` and `README.md` docs accordingly.

## 2. New RHI (Render Hardware Interface) abstraction layer — the headline feature
- Added three **new modules**: `pe-rhi` (abstract rendering API), `pe-rhi-opengl`, and `pe-rhi-vulkan`.
- Introduced abstractions for render pass state: viewport, scissor, input assembly, blend state, depth/stencil, rasterization, multisampling, and a `PrimitiveTopology` replacing old `RenderMode`.
- New **shader abstraction** in RHI with an OpenGL implementation and a **Vulkan implementation** (`lwjgl-shaderc` binding created; old `Shader`/`ShaderProgram` classes in core deprecated/delegated to RHI).
- Added a `VkResult` wrapper and a `VulkanExample`.
- Removed per-module GL state calls (blend/cull/polygon-mode/provoking-vertex/depth) from the LWJGL bindings in favor of the RHI-backed state objects.
- Removed unused PBO and struct `.glsl` helper files.

## 3. Kotlin migration
- Bulk of `pe-core` (and much of the engine/examples) converted from Java → Kotlin (635 `.kt` files in the diff; many `.java → .kt` renames).
- Replaced the `org.reflections` library with `kotlin-reflect`.
- Adopted JOML primitives/custom drawable types, value classes, extension functions, private-set properties.

## 4. Test infrastructure
- Added test conventions in `docs/TESTING.md` (agent testing guidelines).
- Wrote tests for the `pe-math` module.

## 5. Rendering pipeline & DSL
- Render graph **DSL** for render passes, shadows, fog, and screen.
- New renderer/shaders link builder DSL (`ShadersLink`, `RendererPrototype`), unified under `RenderPass`/`Renderer3D`.
- Implemented **SSAO**, **bloom**, and refactored post-processing passes.

+++ pe-gui/src/main/resources/shaders/gui/render/letter.vertex.glsl
@@ -5,7 +5,7 @@ vec2(1, 1), vec2(1, 0)
 );
 
 // Uniforms
-uniform vec4  u_bounds;
+uniform ivec4 u_bounds;
 uniform ivec2 u_bitmapDimensions;
 uniform ivec4 u_bitmapBounds;
 uniform ivec2 u_resolution;
